Added ItemClickListener. Integrated ViewHolder inside the Adapter
