+++ .idea/.name
@@ -1,1 +1,1 @@
-Searchable RecyclerView Demo
+Searchable-RecyclerView-Demo
